chore: Move slimcat importer
